Adding cluster switch from my handmade one to ZooKeeper, this will most likely be the cluster mechanism going forward.
cluster: Started moving toward ZooKeeper

I'm moving it away from my jenk cluster mechanism, and to the builtin
hadoop ZooKeeper one. This is great, as it does all of the behind
the scenes stuff that was necessary for me.

cluster: Add lock and listener code

I've gotten it pretty closely moved over to ZooKeeper, the ring
state is now a shareable object between the nodes. There is also a
global cluster lock that can be used to enforce ordering of events.

cluster: Get the state to update with node join

Nodes may join the cluster, and add their respective addresses and
ports. This does not include their range for hash values, or the
ability to gracefully leave the cluster.

session: Added session state that integrates with ZooKeper

The session object will be a synchronized cluster object within
ZooKeeper. It will maintain items such as the working directory, and
other properties that are relevant for the filesystem.