disk: Split Data and MetaData into two seperate files
For seperation of concerns, the metadata is going to live in its
own seperate file, partition, or device.

Because of this, the metadata can be thought of as its own data
store that will implement HrfsBlockStore. This will allow an
entity to distribute and maintain the metadata for the node aside
from the other operations of the filesystem.