datastore: Asynchronous writes to store
Provide a insertion callback that will provide a future to a written block on
disk. This is done according to a work queue, attached to a pool of worker
threads. If the work queue is full at the time of insertion, the thread that
attempted to enqueue an insertion will perform the Future.